requirement.txt
asgiref==3.4.1
colorama==0.4.4
cssselect2==0.4.1
Django==3.2.5
lxml==4.6.3
Pillow==8.3.1
pytz==2021.1
qrcode==7.1
qrtools==0.0.2
reportlab==3.5.68
sqlparse==0.4.1
svglib==1.1.0
tinycss2==1.1.0
typing-extensions==3.10.0.0
webencodings==0.5.1